session: move plugin.api.useragents
Move `plugin.api.useragents` to `session.http_useragents`
so that the HTTPSession doesn't depend on the plugin API.

Update tests, formatter/linter rules and updater script accordingly.

Next: restore `plugin.api.useragents` module for regular plugin import.
