+++ src/streamlink/session/http_useragents.py
